% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toneR.R
\name{tone}
\alias{tone}
\title{toneR}
\usage{
tone(sound = 1)
}
\arguments{
\item{sound}{A number or name (character) of the sound to be played.
Possible sounds are:
\enumerate{ \item \code{"purge"} }. If \code{sound} does not match any
of the sounds above a random sound will be played.}
}
\description{
Play a sound.
}
\examples{
# Play a "ping" sound
tone()
}
