% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toneR.R
\name{tone}
\alias{tone}
\title{toneR}
\usage{
tone(sound = 1)
}
\arguments{
\item{sound}{A number or name (character) of the tone to be played.
If \code{sound} does not match any of the inbuilt tones, a random tone will
be played.}
}
\description{
Play a tone
}
\examples{
# Play a "ping" sound
tone()
}
